# Conflicts:
#	pom.xml
#	src/main/java/com/babel/vehiclerentingapproval/exceptions/PersonaNotFoundException.java
#	src/main/java/com/babel/vehiclerentingapproval/models/SolicitudRenting.java
#	src/main/java/com/babel/vehiclerentingapproval/models/TipoResultadoSolicitud.java
#	src/main/java/com/babel/vehiclerentingapproval/persistance/database/mappers/PersonaMapper.java
#	src/main/java/com/babel/vehiclerentingapproval/services/CodigoResolucionValidatorImpl.java
#	src/main/java/com/babel/vehiclerentingapproval/services/EmailServiceImpl.java
#	src/main/java/com/babel/vehiclerentingapproval/services/ProfesionServiceImpl.java
#	src/main/java/com/babel/vehiclerentingapproval/services/ResolucionSolicitudesServiceImpl.java
#	src/main/java/com/babel/vehiclerentingapproval/services/SolicitudRentingServiceImpl.java

+++ src/main/java/com/babel/vehiclerentingapproval/services/preautomaticresults/impl/DenyRulesServicesImpl.java
@@ -53,7 +53,7 @@ public class DenyRulesServicesImpl implements DenyRulesService {
         int month = fechaNacimiento.getMonth() + 1;
         LocalDate fechaConcreta = LocalDate.of(anyo, month, day);
         long anios = ChronoUnit.YEARS.between(fechaConcreta, fechaActual);
-        if (anios + solicitudRenting.getPlazo() >= anyosPlazo) {
+        if (anios + solicitudRenting.getPlazo().intValue() >= anyosPlazo) {
             return true;
         } else {
             return false;
